Added button connection to app store, reorganized files
