--- frontend/src/components/AppList/AppList.jsx
@@ -1,9 +1,0 @@
-import React, { Component } from "react";
-
-export default class AppList extends Component {
-    render() {
-        return (
-            <div>This is a good example of where App Cards will me used</div>
-        );
-    }
-}
