+++ frontend/src/pages/AppPage/AppPage.jsx
@@ -1,15 +1,28 @@
 import React, { Component } from "react";
 import { Container, Title, LineBreak, Apps } from "./AppPage.styled";
-import AppStore from "./AppStore";
+import AppStore from "../../components/AppStore/AppStore";
+import { withRouter } from "react-router-dom";
+import PropTypes from "prop-types";
+import { connect } from "react-redux";
 
-export default class AppPage extends Component {
+class AppPage extends Component {
     render() {
         return (
             <Container>
-                <Title>Apps</Title>
+                <Title>App Store</Title>
                 <LineBreak />
                 <AppStore />
             </Container>
         );
     }
 }
+
+AppPage.propTypes = {
+    auth: PropTypes.object.isRequired,
+};
+
+const mapStateToProps = (state) => ({
+    auth: state.auth,
+});
+
+export default connect(mapStateToProps)(withRouter(AppPage));
