+++ frontend/src/utils/AppDetailsExamples.jsx
